Fix mocha tests
