+++ modules/validation/common-react/__tests__/Validation.test.ts
@@ -1,6 +1,5 @@
 // tslint:disable:no-unused-expression
 import { expect } from 'chai';
-import i18next from 'i18next';
 import CommonModule from '@gqlapp/module-common';
 import i18n from '@gqlapp/i18n-common-react';
 import validation from '..';
